Rename pages and move layouts to follow NextJS file structure convention
