+++ c7_reactjs/src/components/footer.jsx
@@ -1,0 +1,7 @@
+export default function Footer() {
+  return (
+    <footer>
+      <p class="text-center text-body-secondary small">© 2025 Company, Inc</p>{" "}
+    </footer>
+  );
+}
