--- c7_reactjs/src/layouts/main-layout.jsx
@@ -1,11 +1,0 @@
-import { Outlet } from "react-router";
-import Header from "../components/header";
-
-export default function MainLayout() {
-  return (
-    <div className="container">
-      <Header />
-      <Outlet />
-    </div>
-  );
-}
