+++ c7_reactjs/src/pages/page.jsx
@@ -1,0 +1,11 @@
+import Hero from "../components/hero";
+
+export default function HomePage() {
+  return (
+    <Hero
+      title="Group Cat"
+      desc="Kucing adalah simbol kekuatan yang sangat mematikan. Berikan aku satu kucing maka akan kuberi whiskas. Berikan aku sepuluh kucing maka akan kucabut semeru dari akarnya."
+      imgSrc="https://fastly.picsum.photos/id/17/2500/1667.jpg?hmac=HD-JrnNUZjFiP2UZQvWcKrgLoC_pc_ouUSWv8kHsJJY"
+    />
+  );
+}
